feat: claim reference for network myriad
